+++ lib/index.js
@@ -1,4 +1,6 @@
-require('./stylesheets/foods.scss');
-require('./objects/foods.js');
-require('./objects/meals');
+require('./handlers')
+require('./foods');
+require('./meals');
+require('./request')
 require('./application.scss');
+require('./stylesheets/foods.scss');
